remove all unused files
