--- lib/simple_service/command.rb
@@ -1,4 +1,0 @@
-module SimpleService
-  module Command
-  end
-end
