--- lib/simple_service/context.rb
@@ -1,32 +1,0 @@
-module SimpleService
-
-  # TODO: move context logic here?
-  module Context
-
-    def success?
-
-    end
-
-    def success!
-
-    end
-
-    def failure?
-
-    end
-
-    def failure!
-
-    end
-
-    def update_getters_and_setters
-
-    end
-
-    def prune!(*keys)
-
-    end
-
-  end
-
-end
