--- spec/lib/ensure_organizer_is_valid_spec.rb
@@ -1,17 +1,0 @@
-require 'spec_helper'
-
-describe SimpleService::EnsureOrganizerIsValid do
-
-  context '#call' do
-
-    class FooCommand < SimpleService::Command
-      def call; true; end
-    end
-
-    class BadInheritanceCommand
-      def call; true; end
-    end
-
-  end
-
-end
